apps/api/requirements.txt
duckdb
python-dateutil
fastapi
uvicorn[standard]
pydantic>=2
pydantic-settings
google-cloud-storage
pandas
numpy
joblib
lightgbm
pyarrow
# (gcsfs est inutile si on lit via google-cloud-storage; optionnel sinon)
